Move the OpenAPI Object into a generic OAS validator
